pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.64.0",
    "openpyxl>=3.1.5",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
    "pywin32>=308",
    "sympy>=1.13.3",
]
